+++ nuxt.config.js
@@ -59,6 +59,15 @@ export default {
     manifest: {
       lang: 'en',
       crossorigin: 'anonymous',
+      share_target: {
+        action: '/share',
+        method: 'GET',
+        params: {
+          title: 'title',
+          text: 'text',
+          url: 'url',
+        },
+      },
     },
   },
 
@@ -69,13 +78,12 @@ export default {
   },
 
   vuetify: {
-    defaultAssets: false,
+    defaultAssets: true,
     theme: {
       themes: {
         light: {
           primary: 'E37C43', // #E53935
-          secondary: colors.red.lighten4, // #FFCDD2
-          accent: colors.indigo.base, // #3F51B5
+          secondary: 'F0EFE8', // #FFCDD2
         },
       },
     },