+++ pages/search.vue
@@ -1,6 +1,0 @@
-<template>
-  <!-- https://vuetifyjs.com/en/components/autocompletes/#api-search -->
-  <div>
-    <search-input></search-input>
-  </div>
-</template>