+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.8.0] - Add Search, Obboarding and Sharing Support
+
+- Add Onboarding
+- Add suggested Blogs to subscribe
+- Add search feeds
+- Share original link to outside
+- Accept share of link to add as feed
+
 ## [0.7.0] - Add Vuetify
 
 - Remove ZuckJs for stories